% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{methdbr_d}
\alias{methdbr_d}
\title{Methylome matrix.}
\format{A matrix of 187 samples x 123110 CpG probes beta values (beta in [0,1], it includes NA values).}
\usage{
methdbr_d
}
\description{
A matrix of 187 x 123110 beta values (methylation).
}
\keyword{datasets}

