% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{methdbr_e}
\alias{methdbr_e}
\title{Samples metadata.}
\format{A data frame with 187 rows and 9 colums.
\describe{
  \item{sex}{The sex of the samples.}
  \item{age_min}{The age of the samples.}
  \item{tissue}{The tissues of the samples.}
  \item{tissue_stage}{The tissue stage of the samples.}
  \item{tissue_group_level1}{A tissue describtion.}
  \item{tissue_group_level2}{A tissue describtion.}
  \item{tissue_group_level3}{A tissue describtion.}
}}
\usage{
methdbr_e
}
\description{
A data frame that describes the 187 samples.
}
\keyword{datasets}

