% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{methdbr_pf}
\alias{methdbr_pf}
\title{Probes metadata.}
\format{A data frame with 123110 rows and 28 colums. The 9 first colums describe the CpG probes, the last columns describe the related genes.
\describe{
  \item{Chromosome}{The probe chromosome.}
  \item{Start}{The probe position.}
  \item{CGI_Coordinate}{The coordinates of the corresponding CpG island}
  \item{Feature_Type}{The type of the feature in the corresponding CpG island.}
  \item{chrom_text}{The gene chromosome.}
  \item{tx_start}{The min of the gene TSS or TES according to strand.}
  \item{tx_end}{The max of the gene TSS or TES according to strand.}
  \item{gene_symbol}{A gene name.}
  \item{entrez}{The ENYTREZ ID of the gene.}
  \item{strand}{A gene strand.}
  \item{id_assembly}{The version of the genome assembly.}
  \item{exon_length_bp}{the cum lenght of the gene exons}
  \item{type}{The trype of the gene}
  \item{tss_cpg_status}{The CpG status of the gene TSS}
  \item{cpg_density}{the density of CpG.}
}}
\usage{
methdbr_pf
}
\description{
A data frame that describes the 123110 CpG probes.
}
\keyword{datasets}

